AVIServer/src/gtfs_rt_generator/requirements.txt
psycopg2==2.5.4
paho-mqtt==1.0
python-dateutil==2.3
pytz==2014.4

